Moved upgrade functions into project
