+++ src/bin/cli.yml
@@ -5,6 +5,15 @@ settings:
 - StrictUtf8
 args:
 subcommands:
+  - info:
+      about: Displays blueprint info for this project
+      args:
+      - dir:
+          short: d
+          long: dir
+          help: The project directory path
+          takes_value: true
+
   - init:
       about: Initializes a project from a blueprint
       args:
@@ -57,6 +66,7 @@ subcommands:
             long: ssh-key
             help: The path to the private SSH key for git auth
             takes_value: true
+
   - check:
       about: Checks whether a generated project can still be safely updated
       args:
@@ -65,6 +75,7 @@ subcommands:
           long: project
           help: The path to the project
           takes_value: true
+
   - update:
       about: Updates a generated project to a new blueprint version
       args:
@@ -79,14 +90,7 @@ subcommands:
           help: The blueprint to update to
           takes_value: true
           required: true
-  - info:
-      about: Displays blueprint info for this project
-      args:
-      - dir:
-          short: d
-          long: dir
-          help: The project directory path
-          takes_value: true
+
   - upgrade:
       about: Upgrade an existing project to latest blueprint version
       args:
